The OpenCL version is now compatible with single and double precision. Just modify the oclparam.h file to set the desired precision (real_t definition).
Note that NVIDIA OpenCL maintains a cache in ~/.nv/ComputeCache which might disable the compilation. Just remove that directory to make sure that the OpenCL build has the same precision as the main binary. It took me quite a while to figure out why the code was crashing with weird diagnostics.