Added the LOOPFORM option to describe an alternate way of doing parallelism.
Use the LOOP=O option on the make command to activate it.

Please note that on line 64 of trace.c the LOOPFORM is not yet available due to a problem in icx version 2021.1 of Intel's OneAPI.

The TEAMSDIS macro expands to teams distribute only for the GPU version (TARGETON defined) for the time being. Further investigations are needed to see if the
teams distribute structure is usefull for a CPU.